backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests

langchain
langchain-community
langchain-google-genai

google-generativeai

httpx